Revert "[-] obsolete files"
This reverts commit 15b62b062e0604e351a19b998544c3b53a5ab722.

+++ drawer.h
@@ -16,7 +16,7 @@ struct Drawer {
 		height(height) {}
 
 	void draw(const MyRectangle * obj) {
-		draw_rect((int)obj->left(), (int)obj->top(), (int)obj->right(), (int)obj->bottom(), (int)obj->color);
+		draw_rect(obj->left(), obj->top(), obj->right(), obj->bottom(), obj->color);
 	}
 
 	int trim(int x, int len) {
